Rrefactor GameHook.h into a few separate headers
